requirements.txt
streamlit
langchain>=0.2.0,<0.3.0
langchain-openai>=0.1.0,<0.2.0
pypdf2
python-dotenv
pydantic>=2.0.0,<3.0.0 